created resetPasswordReducer module
